+++ src/constants.js
@@ -35,6 +35,7 @@ export const itemStatuses = {
   DECLARED_LOST: 'Declared lost',
   WITHDRAWN: 'Withdrawn',
   CLAIMED_RETURNED: 'Claimed returned',
+  LOST_AND_PAID: 'Lost and paid',
 };
 
 export const requestTypesMap = {
@@ -139,3 +140,7 @@ export const requestStatusFilters = [
 ];
 
 export const pickSlipType = 'pick slip';
+
+export const DOMAIN_NAME = 'requests';
+
+export const APP_ICON_NAME = 'requests';